backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
langchain
langchain-groq
langgraph
chromadb
sentence-transformers
paddleocr
paddlepaddle
PyMuPDF
pdfplumber
opencv-python
pillow
python-dotenv
